requirements.txt
flask
sentence-transformers
numpy
scikit-learn
python-dotenv
cryptography
